Dependency Injection for Render Pipeline
Now Function pointers are passed to the Render pipeline these pointers change how a scene is interpreted which will allow you to swap between rendering with just vertex and vertex and textures. The implementation isnt complete and some old code that I plan on replacing soon is hitching a ride in this commit.